Package cu introduced. Renames of classes concerning the architecture sketches
