+++ ls-mandelbrot-core/src/main/java/de/lexasoft/mandelbrot/cu/package-info.java
@@ -1,0 +1,12 @@
+/**
+ * This package contains the calculation unit (cu) of the Mandelbrot set
+ * calculation.
+ * <p>
+ * These are all classes, which are directly involved in the mathematical
+ * calculation.
+ * 
+ * @since 1.0
+ * @author nierax
+ */
+
+package de.lexasoft.mandelbrot.cu;
