+++ ls-mandelbrot-core/src/main/java/de/lexasoft/mandelbrot/dto/package-info.java
@@ -1,0 +1,10 @@
+/**
+ * Input classes, containing the raw information about the wanted calculation.
+ * <p>
+ * In most cases read form a corresponding yaml file.
+ * 
+ * @see de.lexasoft.mandelbrot.dto.DTO2PropertiesMapper
+ * @since 1.0
+ * @author nierax
+ */
+package de.lexasoft.mandelbrot.dto;
