+++ ls-mandelbrot-core/src/test/java/de/lexasoft/mandelbrot/MandelbrotRunnerFactoryTest.java
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 /**
- * @author admin
+ * @author nierax
  *
  */
 class MandelbrotRunnerFactoryTest {
