+++ ls-mandelbrot-core/src/test/java/de/lexasoft/mandelbrot/cu/MandelbrotFormulaTest.java
@@ -1,9 +1,9 @@
 /**
  * 
  */
-package de.lexasoft.mandelbrot;
+package de.lexasoft.mandelbrot.cu;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.stream.Stream;
 
@@ -13,29 +13,31 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 /**
- * @author admin
+ * @author nierax
  *
  */
-class MandelbrotPointTest {
-	
-	private MandelbrotPoint cut;
+class MandelbrotFormulaTest {
+
+	private MandelbrotFormula cut;
 
 	/**
 	 * @throws java.lang.Exception
 	 */
 	@BeforeEach
 	void setUp() throws Exception {
-		cut= new MandelbrotPoint();
+		cut = new MandelbrotFormula();
 	}
-	
+
 	private static Stream<Arguments> testIterate() {
 		return Stream.of(Arguments.of(0.0, 0.0, 500, true), Arguments.of(0.75, -0.75, 500, false));
 	}
 
 	/**
-	 * Test method for {@link de.lexasoft.mandelbrot.MandelbrotPoint#iterate(double, double, int)}.
+	 * Test method for
+	 * {@link de.lexasoft.mandelbrot.cu.MandelbrotFormula#iterate(double, double, int)}.
 	 * <p>
-	 * Checks some known points of the Mandelbrot set to evaluate that the method works correctly.
+	 * Checks some known points of the MandelbrotIterator set to evaluate that the
+	 * method works correctly.
 	 */
 	@ParameterizedTest
 	@MethodSource
